fix flutter/dart wasm

+++ veilid-core/src/storage_manager/inspect_value.rs
@@ -156,7 +156,7 @@ impl StorageManager {
                     // Get number of subkeys from schema and ensure we are getting the
                     // right number of sequence numbers betwen that and what we asked for
                     #[allow(clippy::unnecessary_cast)]
-                    if answer.seqs.len() != descriptor_info.subkeys.len() as usize {
+                    if answer.seqs.len() as u64 != descriptor_info.subkeys.len() as u64 {
                         // Not the right number of sequence numbers
                         // Move to the next node
                         return Ok(NetworkResult::invalid_message(format!(
